requirements.txt
numpy
keyboard
pytesseract
Pillow